requirements.txt
streamlit
google-generativeai
pandas
numpy
PyPDF2
Pillow